requirements.txt
fastapi
uvicorn[standard]
streamlit
anthropic
python-dotenv
requests
pydantic
